requirements.txt
PyQt5
hid
NumPy
PyWinCtl
pynput
PyYAML
platformdirs
geocoder
pandas
pvlib
requests
pyserial